Share the scaffolding around the feature file script tests
Both scripts are exercised the way they are used, by running them over a
directory of feature files written into a temporary directory, so both test
classes carried the same setup: creating and removing that directory,
writing a feature file into it, running the script from within it, and
reading back what it wrote.

The scaffolding moves to `FeatureFilesTestCase`, which the two classes
extend. Each names the script it is about and the flags to run it with, and
the rest follows: the temporary directory is still named after the script,
and the code style check still runs it without loading `php.ini` while the
analysis needs one loaded for ext-tokenizer.

The listing of what extraction wrote now leaves out files that are not PHP
in both, rather than only in the one whose target directory holds a
manifest. Extraction only ever writes `.php` files, so anything else in
there came from somewhere else.
